pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Raffey", email = "muhammadraffey26@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.11",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[project.scripts]
code = "code:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
